Distance Calculator 1.0.1
A function to calculate distance in a value in a community matrix (full combination of all plots/sites/islands...)

focal - First plot Column or Matrix/Data frame with full combination of all plots
target - Second Plot Column (if focal is not a matrix)
var - matrix where first column MUST be plot/site/island and second to n are variables where the distance between to plots should be calculated. 
For geographic (not geosphere) distance, calculate distance x and y & go Pythagoras.
If Factor or character will return if it is "Equal" or which two factors (sorted) are involved.